Remove unneeded static_cast.

+++ src/compton_tools/Compton_Native.cc
@@ -154,38 +154,38 @@ void Compton_Native::broadcast_MPI(int errcode) {
   Insist(data_size < (1U << 31U), "Unrepresentable with int needed in the broadcast");
 
   // Derived sizes
-  const auto tsz = static_cast<int>(num_temperatures_);
-  const auto egsz = static_cast<int>(num_groups_ + 1U);
-  const auto fgsz = static_cast<int>(num_temperatures_ * num_groups_);
-  const auto isz = static_cast<int>(fgsz + 1U);
-  const auto dsz = static_cast<int>(data_size);
+  const auto tsz = num_temperatures_;
+  const auto egsz = num_groups_ + 1U;
+  const auto fgsz = num_temperatures_ * num_groups_;
+  const auto isz = fgsz + 1U;
+  const auto dsz = data_size;
 
   // Broadcast grids
   if (rank != bcast_rank)
     Ts_.resize(tsz);
-  rtt_c4::broadcast((tsz > 0 ? &Ts_[0] : nullptr), tsz, 0);
+  rtt_c4::broadcast((tsz > 0 ? &Ts_[0] : nullptr), tsz, 0u);
 
   if (rank != bcast_rank)
     Egs_.resize(egsz);
-  rtt_c4::broadcast((egsz > 0 ? &Egs_[0] : nullptr), egsz, 0);
+  rtt_c4::broadcast((egsz > 0 ? &Egs_[0] : nullptr), egsz, 0u);
 
   // Broadcast sparse data structures
   if (rank != bcast_rank)
     first_groups_.resize(fgsz);
-  rtt_c4::broadcast((fgsz > 0 ? &first_groups_[0] : nullptr), fgsz, 0);
+  rtt_c4::broadcast((fgsz > 0 ? &first_groups_[0] : nullptr), fgsz, 0u);
 
   if (rank != bcast_rank)
     indexes_.resize(isz);
-  rtt_c4::broadcast((isz > 0 ? &indexes_[0] : nullptr), isz, 0);
+  rtt_c4::broadcast((isz > 0 ? &indexes_[0] : nullptr), isz, 0u);
 
   // Broadcast data itself
   if (rank != bcast_rank)
     data_.resize(dsz);
-  rtt_c4::broadcast((dsz > 0 ? &data_[0] : nullptr), dsz, 0);
+  rtt_c4::broadcast((dsz > 0 ? &data_[0] : nullptr), dsz, 0u);
 
   if (rank != bcast_rank)
     derivs_.resize(dsz);
-  rtt_c4::broadcast((dsz > 0 ? &derivs_[0] : nullptr), dsz, 0);
+  rtt_c4::broadcast((dsz > 0 ? &derivs_[0] : nullptr), dsz, 0u);
 }
 
 //------------------------------------------------------------------------------------------------//
